Fixed bug that when there are no complex vars
When performing the grep on line 95 to look for complex variants, if it didn't find any the grep would return interger(0). So when we subset to remove complex variants with vcf[-complex_vars,] it was removing all of the variants. 

I made this conditional now so that the subsetting only occurs when at least one complex variant was observed.